implements create new post
